+++ docs/.vuepress/utils/alias.js
@@ -2,13 +2,13 @@
  * @Author: Rainy
  * @Date: 2020-02-27 16:38:27
  * @LastEditors: Rainy
- * @LastEditTime: 2020-02-27 18:00:37
+ * @LastEditTime: 2020-06-25 17:36:03
  */
 
 const alias = {
   'guide': '介绍',
-  'how-to-config-docs': '如何配置文档? ',
-  'how-to-write-docs': '如何写文档? ',
+  'what-is-typescript': 'Typescript 是什么? ',
+  'basic': 'Typescript 基础类型 ',
 };
 
 module.exports = {