+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# 1.0.0 (2020-03-04)
+# 1.0.0 (2020-06-26)
 
 
 ### Features
 
-* vuepress docs template ([77f3649](https://github.com/Rain120/vuepress-docs-template/commit/77f36499665edfad4f0333d6a55f54a541177f06))
+* plop auto create docs model ([1f5a2f9](https://github.com/Rain120/typescript-guide/commit/1f5a2f94bbd7a430bbfbb48a09718b8ec9d14190))
+* vuepress docs template ([77f3649](https://github.com/Rain120/typescript-guide/commit/77f36499665edfad4f0333d6a55f54a541177f06))